LayoutTests/storage/indexeddb/detached-iframe-expected.txt

ALERT: original value
PASS successfullyParsed is true

TEST COMPLETE
